fix behaviour in error case
FF sometimes returns undefined for el.innerText (can't test that because
the test environment doesn't seem to do this) - if el.textContent is the
empty string this would result in an exception.

The try-catch is just additional save-guarding - the FormatPreserver
should never throw, and fail just by returning the input.